change double to decimal

+++ Statistic.Application/Views/QuizScoreView.cs
@@ -4,6 +4,6 @@
     {
         public string Title { get; set; }
 
-        public double Score { get; set; }
+        public decimal Score { get; set; }
     }
 }
